+++ src/components/styles/global.css
@@ -1,5 +1,5 @@
 body {
-  background-color: #faf;
+  background-color: #fafafa;
   padding: 0;
   margin: 0;
   box-sizing: border-box;
