+++ tailwind.config.js
@@ -7,7 +7,7 @@ module.exports = {
     extend: {
       fontFamily: {
         sans: ['Roboto', 'sans-serif']
-      } 
+      }, 
     },
   },
   plugins: [],
